fix: move hooks to services folder
