--- src/pages/RepoPage/SettingsTab/tabs/GeneralTab/Tokens/TokensTeam/hooks/index.js
@@ -1,1 +1,0 @@
-export * from './useRepoSettingsTeam'
