--- src/pages/RepoPage/SettingsTab/tabs/GeneralTab/hooks/index.js
@@ -1,1 +1,0 @@
-export * from './useRepoDefaultBranch'
