+++ src/services/repo/index.js
@@ -3,6 +3,8 @@ export * from './useRepo'
 export * from './useRepoOverview'
 export * from './useRepoCoverage'
 export * from './useRepoSettings'
+export * from './useRepoSettingsTeam'
+export * from './useRepoDefaultBranch'
 export * from './useEncodeString'
 export * from './useActivateFlagMeasurements'
 export * from './useEraseRepoContent'
